Remove properties from `InteractionBarConfiguration`
